added structured code
