+++ sedc-server/sedc-server-core/Entities/Request.cs
@@ -1,7 +1,10 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Text;
+using System.Threading.Tasks;
 
-namespace Sedc.Server.Core
+namespace Sedc.Server.Core.Entities
 {
     internal class Request
     {
@@ -30,5 +33,4 @@ namespace Sedc.Server.Core
 ";
         }
     }
-
 }
