+++ sedc-server/sedc-server-core/Entities/Server.cs
@@ -1,8 +1,15 @@
-﻿using System;
+﻿using Sedc.Server.Core.Helpers.RequestHelpers;
+using Sedc.Server.Core.Helpers.ResponseHelpers;
+using Sedc.Server.Core.Helpers.ServerHelpers;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net;
 using System.Net.Sockets;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace Sedc.Server.Core
+namespace Sedc.Server.Core.Entities
 {
     public class Server
     {
@@ -19,7 +26,7 @@ namespace Sedc.Server.Core
         public void Start()
         {
             var address = IPAddress.Loopback;
-            var port = options.Port; 
+            var port = options.Port;
             TcpListener listener = new TcpListener(address, port);
             listener.Start();
             Console.WriteLine($"Started listening on port {port}");
