+++ sedc-server/sedc-server-core/Helpers/RequestHelpers/RequestReader.cs
@@ -1,8 +1,12 @@
-﻿using System;
+﻿using Sedc.Server.Core.Entities;
+using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Text;
+using System.Threading.Tasks;
 
-namespace Sedc.Server.Core
+namespace Sedc.Server.Core.Helpers.RequestHelpers
 {
     internal class RequestReader
     {
