+++ sedc-server/sedc-server-core/Helpers/ResponseHelpers/ResponseSender.cs
@@ -1,8 +1,12 @@
-﻿using System;
+﻿using Sedc.Server.Core.Entities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net.Sockets;
 using System.Text;
+using System.Threading.Tasks;
 
-namespace Sedc.Server.Core
+namespace Sedc.Server.Core.Helpers.ResponseHelpers
 {
     internal class ResponseSender
     {
