+++ sedc-server/sedc-server-core/Helpers/ResponseHelpers/StatusText.cs
@@ -1,19 +1,12 @@
+﻿using Sedc.Server.Core.Enums;
+using System;
 using System.Collections.Generic;
-using System.ComponentModel;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace Sedc.Server.Core
+namespace Sedc.Server.Core.Helpers.ResponseHelpers
 {
-    public enum StatusEnum
-    {
-        OK = 200,
-        ServerError = 500,
-        NotFound = 404,
-        BadRequest = 400,
-        Unauthorized = 401,
-        Forbidden = 403,
-        SemanticError = 422
-    }
-
     public static class StatusText
     {
         private static Dictionary<StatusEnum, string> textStatuses = new();
