+++ sedc-server/sedc-server-try-two/Program.cs
@@ -1,5 +1,6 @@
 ﻿using Sedc.Server.Core;
-
+using Sedc.Server.Core.Entities;
+using Sedc.Server.Core.Helpers.ServerHelpers;
 using System;
 
 namespace sedc_server_try_two
